modify how the application runtime
Instead of a single python file, now there is a perceptron class,
which instansiates a single neuron and a main file that starts the
application. The user now has the option to choose between training
the perceptron, feeding it custom values to get guesses and displaying
the actual weights at that time.

Changes made with this commit:
- Modified perceptron.py to be a class file
- Added main.py file
- Added the ability for the user to train the perceptron
- Added the ability to feed custom values to get guesses
- Added the ability to see weights
- Edited README.md to show more info

Now it is possible to create perceptron instances and play with them.
While they may only be able to output binary values, this opens the
possibility to create a multi-layer perceptron (MLP) with these
instances in the future.