fix movement and reset method
This commit fixes the movement of the agent on the canvas to match
the change in the tables. This is important as the coordinates (x,y)
are not the same in the tables (rows, columns). In addition to this
fix, the issue regarding the reset method is also solved. In general
changes happened to the codebase, deleting unnecessary code.

Resolves #5, resolves #6